fix: rename social.js to socialAPI.js to resolve CI module resolution issue
- Renamed frontend/src/api/social.js to socialAPI.js
- Updated all imports in useSocialAccounts.js and tests
- This resolves the persistent 'Module not found' error in GitHub Actions
- The issue was likely caused by git history confusion from the .ts to .js rename
- All tests pass locally (42/42 Social Panel tests)
- Build completes successfully
