+++ frontend/src/api/__tests__/social.test.js
@@ -4,7 +4,7 @@
  * Tests for the social networks API mock implementation
  */
 
-import socialAPI from '../social';
+import socialAPI from '../socialAPI';
 
 // These tests verify the API SDK functions work correctly 
 // They use the real implementation but in mock mode
